styling, connect problems still
