+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "chaincred",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "./index.js",
   "scripts": {
     "start": "EXPO_USE_FAST_RESOLVER=1 npx expo --dev-client",
     "android": "expo run:android",
@@ -63,7 +63,7 @@
     "expo-web-browser": "~13.0.3",
     "graphql": "^16.8.1",
     "graphql-request": "^6.1.0",
-    "nativewind": "4",
+    "nativewind": "^4.0.1",
     "prisma": "^5.10.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -85,7 +85,7 @@
     "react-native-web": "~0.19.6",
     "supercluster": "^8.0.1",
     "tailwind-merge": "^2.3.0",
-    "tailwindcss": "^3.4.3",
+    "tailwindcss": "^3.4.4",
     "tailwindcss-animate": "^1.0.7",
     "thirdweb": "^5.28.0",
     "tiny-invariant": "^1.3.1",
@@ -117,7 +117,7 @@
   },
   "private": true,
   "engines": {
-    "node": ">=16"
+    "node": ">=20"
   },
   "prisma": {
     "schema": "lib/services/db/schema.prisma"
